planventure-api/requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3

# Authentication & Security
bcrypt==4.0.1
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9  # For PostgreSQL support
alembic==1.12.0  # For database migrations

# Development & Testing
pytest==7.4.2
black==23.9.1
flake8==6.1.0

# Validation
jsonschema==4.19.1

# OAuth and Google API
requests==2.31.0
oauthlib==3.2.2
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.97.0
requests-oauthlib==1.3.1